+++ packages/botonic-plugin-flow-builder/src/webview/contents-context.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react'
+
+import { WebviewContentsContextType } from './types'
+
+export const WebviewContentsContext = createContext<WebviewContentsContextType>(
+  {
+    getTextContent: () => undefined,
+    getImageSrc: () => undefined,
+    setCurrentLocale: () => undefined,
+  }
+)