+++ packages/botonic-react/src/index.ts
@@ -1,6 +1,10 @@
 export { BotonicInputTester, BotonicOutputTester } from './botonic-tester'
 export * from './components'
-export { RequestContext, WebchatContext } from './contexts'
+export {
+  RequestContext,
+  WebchatContext,
+  WebviewRequestContext,
+} from './contexts'
 export { DevApp } from './dev-app'
 export * from './index-types'
 export { msgsToBotonic, msgToBotonic } from './msg-to-botonic'